+++ app/lib/shared/versions.dart
@@ -22,10 +22,10 @@ final RegExp runtimeVersionPattern = RegExp(r'^\d{4}\.\d{2}\.\d{2}$');
 /// when the version switch happens.
 const acceptedRuntimeVersions = <String>[
   // The current [runtimeVersion].
-  '2023.02.14',
+  '2023.02.22',
   // Fallback runtime versions.
+  '2023.02.14',
   '2023.02.10',
-  '2023.02.09',
 ];
 
 /// Represents a combined version of the overall toolchain and processing,
@@ -55,9 +55,9 @@ bool shouldGCVersion(String version) =>
 // keep in-sync with SDK version in .mono_repo.yml and Dockerfile
 final String runtimeSdkVersion = '2.19.0';
 final String toolStableDartSdkVersion = '2.19.2';
-final String toolStableFlutterSdkVersion = '3.7.3';
+final String toolStableFlutterSdkVersion = '3.7.4';
 final String toolPreviewDartSdkVersion = '3.0.0-218.0.dev';
-final String toolPreviewFlutterSdkVersion = '3.7.3';
+final String toolPreviewFlutterSdkVersion = '3.8.0-10.1.pre';
 
 final semanticToolStableDartSdkVersion =
     Version.parse(toolStableDartSdkVersion);